src/co-agent/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.32.0
azure-servicebus==7.14.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.8
h11==0.14.0
idna==3.10
isodate==0.7.2
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
